сделал рефакторинг
