+++ webapp/parsers/utils.py
@@ -1,4 +1,5 @@
 import time
+from datetime import datetime, timedelta
 from selenium import webdriver
 from webapp import db
 from webapp.model import RealEstateAds
@@ -27,4 +28,39 @@ def save_flat(url, title, date, price):
     if not flat_exits:
         new_flat = RealEstateAds(title=title, url=url, date=date, price=price)
         db.session.add(new_flat)
-        db.session.commit()   
+        db.session.commit()
+
+def parser_room(room_str):
+    if '8-комн' in room_str:
+        room = 8
+    elif '7-комн' in room_str:
+        room = 7
+    elif '6-комн' in room_str:
+        room = 6
+    elif '5-комн' in room_str:
+        room = 5
+    elif '4-комн' in room_str:
+        room = 4
+    elif '3-комн' in room_str:
+        room = 3
+    elif '2-комн' in room_str:
+        room = 2
+    elif '1-комн' in room_str:
+        room = 1
+    else:# студия
+        room = 9
+    return room
+
+
+def parse_date(date_str):# парсим строку на вхождение слова сегодня/вчера и меняем на актуальную дату
+    if 'сегодня' in date_str:
+        today = datetime.now()
+        date_str = date_str.replace('сегодня', today.strftime('%d %B %Y'))
+    elif 'вчера' in date_str:
+        yesterday = datetime.now() - timedelta(days=1)
+        date_str = date_str.replace('вчера', yesterday.strftime('%d %B %Y'))
+    try:
+        return datetime.strptime(date_str, '%d %B %Y, %H:%M')
+    except ValueError:
+        this_year = datetime.today().year# добавляем год к дате
+        return datetime.strptime(date_str, '%d %b, %H:%M').replace(year=this_year)
